feat(core): create progress.py
