+++ core/src/apps/bitcoin/sign_tx/progress.py
@@ -111,7 +111,7 @@ class Progress:
 
     def report_init(self) -> None:
         from trezor import workflow
-        from trezor.ui.layouts import bitcoin_progress, coinjoin_progress
+        from trezor.ui.layouts.progress import bitcoin_progress, coinjoin_progress
 
         progress_layout = bitcoin_progress
         if self.is_coinjoin:
