+++ core/src/trezor/ui/layouts/progress.py
@@ -1,0 +1,1 @@
+from .tt_v2.progress import *  # noqa: F401,F403
